# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SocIA Chatbot API
  description: |
    API completa para o chatbot SocIA da Caixa Econômica Federal.
    
    Arquitetura de microsserviços com três módulos principais:
    - **Chat API Gateway** (porta 3000): Ponto de entrada principal
    - **NLP Processing Service** (porta 3001): Processamento de linguagem natural
    - **Knowledge Base Service** (porta 3002): Base de conhecimento hierárquica
    
    ## Escopo
    - Programas Sociais do Governo Federal
    - Serviços da Caixa Econômica Federal
    - Produtos da Caixa Econômica Federal
  version: 1.0.0
  contact:
    name: SocIA Team
    email: socia@caixa.gov.br
servers:
  - url: http://localhost:3000
    description: Chat API Gateway (Development)
  - url: http://localhost:3001
    description: NLP Processing Service (Development)
  - url: http://localhost:3002
    description: Knowledge Base Service (Development)
...
paths:
...
  /api/nlp/process:
    post:
      tags:
        - NLP
      summary: Processar intenção da mensagem
      description: Extrai intenção, entidades e valida escopo da mensagem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                context:
                  type: object
      responses:
        '200':
          description: Processamento concluído
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  inScope:
                    type: boolean
                  intent:
                    type: object
                    properties:
                      originalMessage:
                        type: string
                      intent:
                        type: string
                      category:
                        type: string
                      topic:
                        type: string
                      confidence:
                        type: number
                      entities:
                        type: object
                      knowledge:
                        type: object
  /api/nlp/validate-scope:
    post:
      tags:
        - NLP
      summary: Validar escopo da mensagem
      description: Verifica se a mensagem está dentro do escopo da SocIA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
      responses:
        '200':
          description: Validação concluída
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      inScope:
                        type: boolean
                      response:
                        type: string
                        nullable: true
...